+++ tests/external_urllib_request/test_requests.py
@@ -12,7 +12,7 @@ from typing import Any
 import pytest
 from wrapture import Event, Tape, instrumentation, timeline
 
-from tests.external_urllib_request.server import Server
+from tests.httpserver import Server
 from wrapture_instrumentation.external_urllib_request import UrllibInstrumentation
 
 OPEN = "urllib.request:OpenerDirector.open"